+++ src/Builder/UnitBuilder.hpp
@@ -19,9 +19,9 @@ public:
 
     void setWorkingModule(icode::ModuleDescription* moduleDescription);
 
-    Unit unitFromIntLiteral(int value);
+    Unit unitFromIntLiteral(long value);
     Unit unitFromCharLiteral(char value);
-    Unit unitFromFloatLiteral(float value);
+    Unit unitFromFloatLiteral(double value);
     Unit unitFromTypeDescription(icode::TypeDescription& typeDescription, const std::string& name);
     Unit unitFromEnum(int enumValue);
     Unit unitFromUnitList(const std::vector<Unit>& unitList);
